[CM_BE_20] Extracting isRequestPresent method to common
