+++ src/main/java/in/projecteka/consentmanager/link/discovery/DiscoveryRepository.java
@@ -1,11 +1,10 @@
 package in.projecteka.consentmanager.link.discovery;
 
+import in.projecteka.consentmanager.common.DbOperation;
 import in.projecteka.consentmanager.common.DbOperationError;
 import io.vertx.pgclient.PgPool;
 import io.vertx.sqlclient.Tuple;
 import reactor.core.publisher.Mono;
-
-import static in.projecteka.consentmanager.clients.ClientError.unknownErrorOccurred;
 
 public class DiscoveryRepository {
 
@@ -33,20 +32,6 @@ public class DiscoveryRepository {
     }
 
     public Mono<Boolean> isRequestPresent(String requestId) {
-        return Mono.create(monoSink ->
-                dbClient.preparedQuery(CHECK_REQUEST_ID_EXISTS)
-                        .execute(Tuple.of(requestId),
-                                handler -> {
-                                    if (handler.failed()) {
-                                        monoSink.error(new DbOperationError());
-                                        return;
-                                    }
-                                    var iterator = handler.result().iterator();
-                                    if (!iterator.hasNext()) {
-                                        monoSink.error(unknownErrorOccurred());
-                                        return;
-                                    }
-                                    monoSink.success(iterator.next().getBoolean(0));
-                                }));
+        return DbOperation.getBooleanMono(requestId, dbClient, CHECK_REQUEST_ID_EXISTS);
     }
 }
