+++ src/main/java/in/projecteka/consentmanager/link/link/LinkController.java
@@ -15,8 +15,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
-import java.util.UUID;
-
 @RestController
 @AllArgsConstructor
 public class LinkController {
