+++ src/main/java/in/projecteka/consentmanager/user/TransactionPinRepository.java
@@ -1,7 +1,7 @@
 package in.projecteka.consentmanager.user;
 
 import in.projecteka.consentmanager.clients.ClientError;
-import in.projecteka.consentmanager.common.DbOperationError;
+import in.projecteka.consentmanager.common.DbOperation;
 import in.projecteka.consentmanager.user.model.TransactionPin;
 import io.vertx.pgclient.PgPool;
 import io.vertx.sqlclient.Row;
@@ -11,7 +11,6 @@ import reactor.core.publisher.Mono;
 
 import java.util.Optional;
 
-import static in.projecteka.consentmanager.clients.ClientError.unknownErrorOccurred;
 
 @AllArgsConstructor
 public class TransactionPinRepository {
@@ -52,21 +51,7 @@ public class TransactionPinRepository {
                         }));
     }
     public Mono<Boolean> isRequestPresent(String requestId) {
-        return Mono.create(monoSink ->
-                dbClient.preparedQuery(CHECK_REQUEST_ID_EXISTS)
-                        .execute(Tuple.of(requestId),
-                                handler -> {
-                                    if (handler.failed()) {
-                                        monoSink.error(new DbOperationError());
-                                        return;
-                                    }
-                                    var iterator = handler.result().iterator();
-                                    if (!iterator.hasNext()) {
-                                        monoSink.error(unknownErrorOccurred());
-                                        return;
-                                    }
-                                    monoSink.success(iterator.next().getBoolean(0));
-                                }));
+        return DbOperation.getBooleanMono(requestId, dbClient, CHECK_REQUEST_ID_EXISTS);
     }
 
     private Optional<TransactionPin> transactionPinFrom(Row row) {
